Gift Card now asks for amount
Before the whole payment would just be clear. It now asks for the amount the giftcard is worth. If it is equal to the total it pays normally. If it only pays partially it subtracts the amount and then presents the user the option to pay again with the new total